rewritten - env.sh files created in every project - menu tilte get now updated on every build in ubuntu - moved to latest ubuntu lts versions: 24.04.1 + 22.24.5

+++ ubuntu22/99_cleanup.sh
@@ -1,10 +1,9 @@
 #!/bin/bash
 
 
-BASEDIR="/srv/iso.remaster/ubuntu22"
-ISO_SRC="$BASEDIR/ubuntu-22.04.4-live-server-amd64.iso"
-ISO_DST="$BASEDIR/ubuntu-22.04-autoinstall-amd64.iso"
-WORK_DIR="$BASEDIR/build"
+test -f env.sh || echo ERROR: env.sh is missing
+test -f env.sh || exit 1
+source env.sh
 
 cd $BASEDIR || exit 1
 rm -rfv $WORK_DIR $BASEDIR/remaster.sh.hint $ISO_DST $ISO_SRC
