+++ ubuntu24/99_cleanup.sh
@@ -1,10 +1,9 @@
 #!/bin/bash
 
 
-BASEDIR="/srv/iso.remaster/ubuntu24"
-ISO_SRC="$BASEDIR/ubuntu-24.04-live-server-amd64.iso"
-ISO_DST="$BASEDIR/ubuntu-24.04-autoinstall-amd64.iso"
-WORK_DIR="$BASEDIR/build"
+test -f env.sh || echo ERROR: env.sh is missing
+test -f env.sh || exit 1
+source env.sh
 
 cd $BASEDIR || exit 1
 rm -rfv $WORK_DIR $BASEDIR/remaster.sh.hint $ISO_DST $ISO_SRC
